Fix base64 encoding

+++ src/wasm2js.h
@@ -448,9 +448,9 @@ static std::string base64Encode(std::vector<char> &data) {
 
   while (i + 3 <= data.size()) {
     int bits =
-      (((int) data[i + 0]) << 16) |
-      (((int) data[i + 1]) << 8) |
-      (((int) data[i + 2]) << 0);
+      (((uint8_t) data[i + 0]) << 16) |
+      (((uint8_t) data[i + 1]) << 8) |
+      (((uint8_t) data[i + 2]) << 0);
     ret += alphabet[(bits >> 18) & 0x3f];
     ret += alphabet[(bits >> 12) & 0x3f];
     ret += alphabet[(bits >> 6) & 0x3f];
@@ -460,14 +460,14 @@ static std::string base64Encode(std::vector<char> &data) {
 
   if (i + 2 == data.size()) {
     int bits =
-      (((int) data[i + 0]) << 8) |
-      (((int) data[i + 1]) << 0);
+      (((uint8_t) data[i + 0]) << 8) |
+      (((uint8_t) data[i + 1]) << 0);
     ret += alphabet[(bits >> 10) & 0x3f];
     ret += alphabet[(bits >> 4) & 0x3f];
     ret += alphabet[(bits << 2) & 0x3f];
     ret += '=';
   } else if (i + 1 == data.size()) {
-    int bits =(int) data[i + 0];
+    int bits = (uint8_t) data[i + 0];
     ret += alphabet[(bits >> 2) & 0x3f];
     ret += alphabet[(bits << 4) & 0x3f];
     ret += '=';
